src/main/java/one/stoorx/leetcode/algorithms/util/ListNode.java
package one.stoorx.leetcode.algorithms.util;

import java.util.Objects;

public class ListNode {
    public int val;
    public ListNode next = null;

    public ListNode(int val) {
        this.val = val;
    }

    public ListNode(int val, ListNode next) {
        this.val = val;
        this.next = next;
    }

    public static ListNode fromArray(int[] array) {
        if (array != null)
            return convertRecursive(array, 0);
        else return null;
    }

    private static ListNode convertRecursive(int[] array, int i) {
        if (i < array.length) {
            return new ListNode(array[i], convertRecursive(array, i + 1));
        } else return null;
    }

    @Override
    public boolean equals(Object o) {
        if (this == o) return true;
        if (o == null || getClass() != o.getClass()) return false;
        ListNode listNode = (ListNode) o;
        return val == listNode.val && Objects.equals(next, listNode.next);
    }

    @Override
    public int hashCode() {
        return Objects.hash(val, next);
    }

    @Override
    public String toString() {
        return "ListNode{" +
                "val=" + val +
                ", next=" + next +
                '}';
    }
}